+++ net/discovery/src/main/java/de/danoeh/antennapod/net/discovery/PodcastSearcherRegistry.java
@@ -21,6 +21,7 @@ public class PodcastSearcherRegistry {
             searchProviders.add(new SearcherInfo(new AudiothekPodcastSearcher(), 1.0f));
             searchProviders.add(new SearcherInfo(new BBCSoundsPodcastSearcher(), 1.0f));
             searchProviders.add(new SearcherInfo(new SRFPlayPodcastSearcher(), 1.0f));
+            searchProviders.add(new SearcherInfo(new ORFSoundPodcastSearcher(), 1.0f));
         }
         return searchProviders;
     }